+++ templates/expo-template-tabs/package.json
@@ -31,9 +31,9 @@
     "react-native-web": "~0.11.7",
     "@react-native-community/masked-view": "~0.1.6",
     "react-native-safe-area-context": "~0.6.4",
-    "@react-navigation/native": "5.0.0-alpha.35",
-    "@react-navigation/stack": "5.0.0-alpha.68",
-    "@react-navigation/bottom-tabs": "5.0.0-alpha.43"
+    "@react-navigation/native": "^5.0.0",
+    "@react-navigation/stack": "^5.0.0",
+    "@react-navigation/bottom-tabs": "^5.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
